@@ -20,37 +20,30 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.2.10.RELEASE</version>
+        <version>2.7.0</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <skipTests>true</skipTests>
+        <spring-boot.version>2.7.0</spring-boot.version>
         <docker.image.prefix>1.117.230.132:5000</docker.image.prefix>
         <dockerfile.maven.plugin.version>1.4.10</dockerfile.maven.plugin.version>
-        <activiti-dependencies.version>7.1.0-M11</activiti-dependencies.version>
-        <druid.version>1.1.23</druid.version>
-        <hutool.version>5.6.2</hutool.version>
-        <guava.version>30.1.1-jre</guava.version>
+        <activiti.version>7.4.1</activiti.version>
+        <druid.version>1.2.9</druid.version>
+        <hutool.version>5.8.0</hutool.version>
+        <guava.version>31.1-jre</guava.version>
         <swagger3.version>3.0.0</swagger3.version>
-        <pagehelper-starter.version>1.3.0</pagehelper-starter.version>
-        <pagehelper.version>5.2.0</pagehelper.version>
-        <mybatis-generator.version>1.4.0</mybatis-generator.version>
-        <mybatis.version>3.5.5</mybatis.version>
-        <mybatis-plus.version>3.4.0</mybatis-plus.version>
-        <velocity.version>2.2</velocity.version>
-        <mysql-connector.version>8.0.20</mysql-connector.version>
-        <sqlserver-connector.version>6.4.0.jre8</sqlserver-connector.version>
-        <spring-data-commons.version>2.3.0.RELEASE</spring-data-commons.version>
+        <mybatis-plus.version>3.5.2</mybatis-plus.version>
         <jjwt.version>0.9.0</jjwt.version>
-        <aliyun-oss.version>2.5.0</aliyun-oss.version>
-        <logstash-logback.version>5.3</logstash-logback.version>
-        <minio.version>7.1.0</minio.version>
-        <gson.version>2.8.5</gson.version>
-        <weixin-java.version>4.0.0</weixin-java.version>
+        <aliyun-oss.version>3.15.0</aliyun-oss.version>
+        <logstash-logback.version>7.2</logstash-logback.version>
+        <minio.version>8.4.1</minio.version>
+        <gson.version>2.9.0</gson.version>
+        <weixin-java.version>4.3.0</weixin-java.version>
         <kaptcha.version>0.0.9</kaptcha.version>
         <!--系统子模块-->
         <common.version>1.0-SNAPSHOT</common.version>
@@ -284,10 +219,9 @@
                 <plugin>
                     <groupId>org.apache.maven.plugins</groupId>
                     <artifactId>maven-compiler-plugin</artifactId>
-                    <version>3.8.1</version>
                     <configuration>
-                        <source>1.8</source>
-                        <target>1.8</target>
+                        <source>17</source>
+                        <target>17</target>
                         <encoding>UTF-8</encoding>
                     </configuration>
                 </plugin>
