@@ -1,10 +1,9 @@
-import com.youjun.api.DataApiApplication;
 import org.activiti.engine.*;
 import org.activiti.engine.impl.cfg.StandaloneProcessEngineConfiguration;
 import org.junit.jupiter.api.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 /**
  * <p>
@@ -14,8 +13,8 @@ import org.springframework.test.context.junit4.SpringRunner;
  * @author kirk
  * @since 2021/11/29
  */
-@SpringBootTest(classes = DataApiApplication.class,webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class ActivitiTest {
 
     /**
