@@ -1,17 +1,16 @@
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.youjun.api.DataApiApplication;
 import com.youjun.api.modules.office.controller.dto.BctDataDTO;
 import com.youjun.common.util.JsonUtils;
 import lombok.extern.slf4j.Slf4j;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.amqp.core.AmqpTemplate;
 import org.springframework.amqp.core.Message;
 import org.springframework.amqp.core.MessageBuilder;
 import org.springframework.amqp.core.MessageDeliveryMode;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.util.HashMap;
 
@@ -24,8 +23,8 @@ import java.util.HashMap;
  * @since 2021/8/7
  */
 @Slf4j
-@SpringBootTest(classes = DataApiApplication.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class RabbitMQTest {
     @Autowired
     private AmqpTemplate amqpTemplate;
