@@ -17,30 +17,74 @@ import org.springframework.stereotype.Component;
 public class SpringContextUtils implements ApplicationContextAware {
 	public static ApplicationContext applicationContext;
 
+	/**
+	 * 获取applicationContext
+	 * @return
+	 */
+	public static ApplicationContext getApplicationContext() {
+		return applicationContext;
+	}
+
 	@Override
 	public void setApplicationContext(ApplicationContext applicationContext)
 			throws BeansException {
 		SpringContextUtils.applicationContext = applicationContext;
 	}
 
+	/**
+	 * 通过name获取Bean
+	 * @param name
+	 * @return
+	 */
 	public static Object getBean(String name) {
 		return applicationContext.getBean(name);
 	}
 
-	public static <T> T getBean(String name, Class<T> requiredType) {
-		return applicationContext.getBean(name, requiredType);
+	/**
+	 * 通过class获取Bean
+	 * @param clazz
+	 * @return
+	 * @param <T>
+	 */
+	public static <T> T getBean(Class<T> clazz) {
+		return applicationContext.getBean(clazz);
 	}
 
+	/**
+	 * 通过name,以及Clazz返回指定的Bean
+	 * @param name
+	 * @param clazz
+	 * @return
+	 * @param <T>
+	 */
+	public static <T> T getBean(String name, Class<T> clazz) {
+		return applicationContext.getBean(name, clazz);
+	}
+
+	/**
+	 * 是否包含该bean
+	 * @param name
+	 * @return
+	 */
 	public static boolean containsBean(String name) {
 		return applicationContext.containsBean(name);
 	}
 
+	/**
+	 * 是否单例
+	 * @param name
+	 * @return
+	 */
 	public static boolean isSingleton(String name) {
 		return applicationContext.isSingleton(name);
 	}
 
-	public static Class<? extends Object> getType(String name) {
-		return applicationContext.getType(name);
+	/**
+	 * 根据bean名称获取该bean的class
+	 * @param beanName
+	 * @return
+	 */
+	public static Class<? extends Object> getType(String beanName) {
+		return applicationContext.getType(beanName);
 	}
-
 }
