@@ -87,8 +87,9 @@ public class SchedulerConfig implements SchedulingConfigurer {
             prop.put("org.quartz.threadPool.class", "org.quartz.simpl.SimpleThreadPool");
             prop.put("org.quartz.threadPool.threadCount", "20");
             prop.put("org.quartz.threadPool.threadPriority", "5");
-            //JobStore配置
-            prop.put("org.quartz.jobStore.class", "org.quartz.impl.jdbcjobstore.JobStoreTX");
+            //JobStore配置 springboot>2.5.6后使用 LocalDataSourceJobStore
+            //prop.put("org.quartz.jobStore.class", "org.quartz.impl.jdbcjobstore.JobStoreTX");
+            prop.put("org.quartz.jobStore.class", "org.springframework.scheduling.quartz.LocalDataSourceJobStore");
             //集群配置
             prop.put("org.quartz.jobStore.isClustered", "true");
             prop.put("org.quartz.jobStore.clusterCheckinInterval", "15000");
