@@ -10,7 +10,8 @@ import com.aliyun.oss.model.OSSObject;
 import com.aliyun.oss.model.PutObjectRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.stereotype.Component;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -18,7 +19,8 @@ import java.io.InputStream;
 import java.net.URL;
 import java.util.*;
 
-@Service
+@Component
+@ConditionalOnProperty(name = "aliyun.enabled", havingValue = "true")
 @Slf4j
 public class AliyunOssUtils {
 
@@ -33,11 +35,12 @@ public class AliyunOssUtils {
 
     /**
      * 单个文件上传
+     *
      * @param fileInputStream 文件流
-     * @param filePath 文件名称
+     * @param filePath        文件名称
      * @return key
      */
-    public String uploadFile(InputStream fileInputStream, String filePath){
+    public String uploadFile(InputStream fileInputStream, String filePath) {
         log.info("Calling uploadFile with the file path {}", filePath);
         // 创建OSSClient实例
         log.info("Connecting to the OSS server.");
@@ -69,12 +72,13 @@ public class AliyunOssUtils {
 
     /**
      * 文件上传
+     *
      * @param fileInputStream 文件流
-     * @param folder 文件夹名称
-     * @param filePath 文件名称
+     * @param folder          文件夹名称
+     * @param filePath        文件名称
      * @return key
      */
-    public String uploadFile(InputStream fileInputStream,String folder,String filePath){
+    public String uploadFile(InputStream fileInputStream, String folder, String filePath) {
         log.info("Calling uploadFile with the file path {}", filePath);
         // 创建OSSClient实例
         log.info("Connecting to the OSS server.");
@@ -106,11 +110,12 @@ public class AliyunOssUtils {
 
     /**
      * 文件上传
+     *
      * @param fileInputStream 文件流
-     * @param filePath 文件名称
+     * @param filePath        文件名称
      * @return key
      */
-    public String uploadFile(FileInputStream fileInputStream, String filePath, OSS ossClient){
+    public String uploadFile(FileInputStream fileInputStream, String filePath, OSS ossClient) {
         filePath = UUID.randomUUID().toString().replaceAll("/|-", "") + File.separator + filePath;
         // 创建PutObjectRequest对象
         PutObjectRequest putObjectRequest = new PutObjectRequest(bucketName, filePath, fileInputStream);
@@ -122,20 +127,21 @@ public class AliyunOssUtils {
 
     /**
      * 批量上传文件
+     *
      * @param fileInputStreams
      * @param filePathList
      * @return
      */
-    public List<String> uploadFileList(List<FileInputStream> fileInputStreams, List<String> filePathList){
+    public List<String> uploadFileList(List<FileInputStream> fileInputStreams, List<String> filePathList) {
         // 创建OSSClient实例
         OSS ossClient = new OSSClientBuilder().build(endPoint, accessKeyId, accessKeySecret);
         List<String> res = new ArrayList<>();
-        for (int i = 0; i < fileInputStreams.size(); i++){
+        for (int i = 0; i < fileInputStreams.size(); i++) {
             String filePath = this.uploadFile(fileInputStreams.get(i), filePathList.get(i), ossClient);
             res.add(filePath);
         }
         // 关闭OSSClient
-        if (ossClient!=null){
+        if (ossClient != null) {
             ossClient.shutdown();
         }
         return res;
@@ -144,22 +150,23 @@ public class AliyunOssUtils {
 
     /**
      * 获取文件流
+     *
      * @param filePath
      * @return
      */
-    public InputStream downloadFile(String filePath){
+    public InputStream downloadFile(String filePath) {
         // 创建OSSClient实例
         OSS ossClient = new OSSClientBuilder().build(endPoint, accessKeyId, accessKeySecret);
         boolean exists = ossClient.doesObjectExist(bucketName, filePath);
-        if (exists){
+        if (exists) {
 
             // 获取文件流
             OSSObject ossObject = ossClient.getObject(bucketName, filePath);
-            if (ossObject != null && ossObject.getObjectContent() != null){
+            if (ossObject != null && ossObject.getObjectContent() != null) {
                 ossClient.shutdown();
                 return ossObject.getObjectContent();
             }
-        }else {
+        } else {
             ossClient.shutdown();
             return null;
         }
@@ -169,20 +176,21 @@ public class AliyunOssUtils {
 
     /**
      * 下载文件到本地
+     *
      * @param filePath
      * @return
      */
-    public boolean downloadFile(String filePath, String key){
+    public boolean downloadFile(String filePath, String key) {
         // 创建OSSClient实例
         OSS ossClient = new OSSClientBuilder().build(endPoint, accessKeyId, accessKeySecret);
         boolean exists = ossClient.doesObjectExist(bucketName, key);
-        if (exists){
+        if (exists) {
             File file = new File(filePath);
             // 下载文件到本地
-            ossClient.getObject(new GetObjectRequest(bucketName,key), file);
+            ossClient.getObject(new GetObjectRequest(bucketName, key), file);
             ossClient.shutdown();
             return true;
-        }else {
+        } else {
             ossClient.shutdown();
             return false;
         }
@@ -191,15 +199,16 @@ public class AliyunOssUtils {
 
     /**
      * 获取访问链接
+     *
      * @param filePath
      * @return
      */
-    public String getFileUrl(String filePath){
+    public String getFileUrl(String filePath) {
         // 创建OSSClient实例
         OSS ossClient = new OSSClientBuilder().build(endPoint, accessKeyId, accessKeySecret);
         boolean exists = ossClient.doesObjectExist(bucketName, filePath);
         String url = null;
-        if (exists){
+        if (exists) {
             // 设置过期时间
             Date expiration = new Date(new Date().getTime() + 3600 * 1000);
             GeneratePresignedUrlRequest request = new GeneratePresignedUrlRequest(bucketName, filePath);
@@ -212,14 +221,15 @@ public class AliyunOssUtils {
 
     /**
      * 获取访问链接
+     *
      * @param filePath
      * @return
      */
-    public String getFileUrl(String filePath,OSS ossClient){
+    public String getFileUrl(String filePath, OSS ossClient) {
         // 创建OSSClient实例
         boolean exists = ossClient.doesObjectExist(bucketName, filePath);
         String url = null;
-        if (exists){
+        if (exists) {
             // 设置过期时间
             Date expiration = new Date(new Date().getTime() + 3600 * 1000);
             GeneratePresignedUrlRequest request = new GeneratePresignedUrlRequest(bucketName, filePath);
@@ -230,18 +240,19 @@ public class AliyunOssUtils {
 
     /**
      * 获取访问链接列表
+     *
      * @param filePathList
      * @return
      */
-    public List<String> getFileUrlList(List<String> filePathList){
+    public List<String> getFileUrlList(List<String> filePathList) {
         List<String> urlList = new ArrayList<>();
         OSS ossClient = new OSSClientBuilder().build(endPoint, accessKeyId, accessKeySecret);
-        if (filePathList != null && filePathList.size() > 0 ){
-            for (String str : filePathList){
+        if (filePathList != null && filePathList.size() > 0) {
+            for (String str : filePathList) {
                 urlList.add(getFileUrl(str, ossClient));
             }
         }
-        if (ossClient != null){
+        if (ossClient != null) {
             ossClient.shutdown();
         }
         return urlList;
@@ -249,18 +260,19 @@ public class AliyunOssUtils {
 
     /**
      * 获取访问链接列表
+     *
      * @param filePathList
      * @return
      */
-    public Map<String, String> getFileUrlMap(List<String> filePathList){
-        Map<String,String> urlList = new HashMap<>();
+    public Map<String, String> getFileUrlMap(List<String> filePathList) {
+        Map<String, String> urlList = new HashMap<>();
         OSS ossClient = new OSSClientBuilder().build(endPoint, accessKeyId, accessKeySecret);
-        if (filePathList != null && filePathList.size() > 0 ){
-            for (String str : filePathList){
-                urlList.put(str,getFileUrl(str, ossClient));
+        if (filePathList != null && filePathList.size() > 0) {
+            for (String str : filePathList) {
+                urlList.put(str, getFileUrl(str, ossClient));
             }
         }
-        if (ossClient != null){
+        if (ossClient != null) {
             ossClient.shutdown();
         }
         return urlList;
@@ -269,18 +281,19 @@ public class AliyunOssUtils {
 
     /**
      * 获取图片缩放图列表
+     *
      * @param filePathList
      * @return
      */
-    public List<String> getPhotoUrlList(List<String> filePathList){
+    public List<String> getPhotoUrlList(List<String> filePathList) {
         List<String> urlList = new ArrayList<>();
         OSS ossClient = new OSSClientBuilder().build(endPoint, accessKeyId, accessKeySecret);
-        if (filePathList != null && filePathList.size() > 0 ){
-            for (String str : filePathList){
+        if (filePathList != null && filePathList.size() > 0) {
+            for (String str : filePathList) {
                 urlList.add(getPhotoUrl(str, ossClient));
             }
         }
-        if (ossClient != null){
+        if (ossClient != null) {
             ossClient.shutdown();
         }
         return urlList;
@@ -289,10 +302,11 @@ public class AliyunOssUtils {
 
     /**
      * 获取处理后到照片
+     *
      * @param key
      * @return
      */
-    public String getPhotoUrl(String key, OSS ossClient){
+    public String getPhotoUrl(String key, OSS ossClient) {
         boolean exists = ossClient.doesObjectExist(bucketName, key);
         String url = null;
         if (exists) {
@@ -311,13 +325,13 @@ public class AliyunOssUtils {
     }
 
 
-
     /**
      * 获取处理后到照片
+     *
      * @param key
      * @return
      */
-    public String getPhotoUrl(String key){
+    public String getPhotoUrl(String key) {
         OSS ossClient = new OSSClientBuilder().build(endPoint, accessKeyId, accessKeySecret);
 
         boolean exists = ossClient.doesObjectExist(bucketName, key);
@@ -334,12 +348,11 @@ public class AliyunOssUtils {
             url = signUrl.toString();
         }
 
-        if (ossClient != null){
+        if (ossClient != null) {
             ossClient.shutdown();
         }
         return url;
     }
-
 
 
 }
